+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.4.11
+- Remove asynchronous support totally from running mod code with `BrowserClient`, as it introduces various problems with NodeJS VM
+- Terminal commands `start` and `stop` in `BrowserClient` should work as expected now
 ## 1.4.10
 - Fix mod context and `BrowserClient#execute()` doesn't share the same context during mod lifetime
 ## 1.4.9-alpha6